# doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "EVE Online Data Warehouse",
    "version": "1.0"
  },
  "produces": ["application/json", "text/plain"],
  "host": "eve-crest-dwh.herokuapp.com",
  "basePath": "/",
  "definitions": {
    "responseError": {
      "type": "object",
      "properties": {
        "errorDetails": {
          "description": "this really should be rationalized but atm, it could be a string OR an object. Swagger doesn't support this as a schema so it's left undefined"
        }
      }
    }
  },
  "securityDefinitions": {
      "swagger-ui-key": {
        "description": "This just defines what the Swagger-UI does so it has to be supported.",
        "name": "api_key",
        "type": "apiKey",
        "in": "query"
      }
  },
  "security": [
    {
      "swagger-ui-key": []
    }
  ],
  "responses": {
    "Error": {
      "description": "Any kind of error, not just an internal error",
      "schema": {
        "$ref": "#/definitions/responseError"
      }
    },
    "Empty": {
      "description": "It's a normal empty response, but swagger can't understand that",
      "schema": {
        "type": "string",
        "enum": [ "" ]
      }
    },
    "NotFound": {
      "description": "Couldn't find something you specified",
      "schema": {
        "type": "string"
      }
    },
    "genericArray": {
      "description": "something",
      "schema": {
        "type": "array",
        "items": {
          "type": "object"
        }
      }
    }
  },
  "paths": {
    "/_/metrics": {
      "get": {
        "responses": {
          "200": {
            "description": "The current metrics",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/v1/types/autocomplete": {
      "get": {
        "description": "fetch details about the type",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "the string to match on",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" }
        }
      }
    },
    "/api/v1/types/{type_id}/details": {
      "get": {
        "description": "fetch details about the type",
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "description": "type_id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" }
        }
      }
    },
    "/api/v1/types/{type_id}/market/stats": {
      "get": {
        "description": "fetch market stats",
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "description": "type_id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "region_id",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "station_id",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "columns",
            "in": "query",
            "description": "columns to return",
            "required": true,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "csv"
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "200": { "$ref": "#/responses/genericArray" }
        }
      }
    }
  }
}